requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arxiv==2.1.3
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
cssselect2==0.7.0
dataclasses-json==0.6.7
distro==1.9.0
docopt==0.6.2
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.8
feedparser==6.0.11
filelock==3.17.0
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gpt-researcher==0.11.0
groq==0.16.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
json_repair==0.36.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-ollama==0.2.3
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.8
loguru==0.7.3
lxml==5.3.1
lxml_html_clean==0.4.1
Markdown==3.7
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
md2pdf==1.0.1
mistune==3.1.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.4.7
openai==1.63.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.1.0
portalocker==2.10.1
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydyf==0.11.0
PyMuPDF==1.25.2
pypdf==5.3.0
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.8
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.1
PyYAML==6.0.2
qdrant-client==1.13.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.20
unstructured-client==0.30.0
urllib3==2.3.0
uvicorn==0.34.0
weasyprint==64.0
webencodings==0.5.1
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zopfli==0.2.3.post1
zstandard==0.23.0
# Real-time audio streaming dependencies
livekit==0.9.2  # Official Python SDK for LiveKit
# Removing faster-whisper with specific version to avoid tokenizer conflicts
faster-whisper>=0.9.0  # Using more flexible versioning 
pyjwt==2.8.0  # JWT handling for authentication
pydub==0.25.1  # Audio processing library
soundfile==0.12.1  # Sound file handling
srt==3.5.3  # Subtitle handling
# Removing kokoro-tts as it seems unavailable - will use another TTS library
gTTS==2.5.0  # Google Text-to-Speech library
